refactor (DiscountPolicies) : 할인 헤택이 0원이라면 혜택에 포함되지 않도록 리팩토링 한다

+++ src/main/java/christmas/domain/discount/WeekendDiscountPolicy.java
@@ -27,7 +27,7 @@ public class WeekendDiscountPolicy implements DiscountPolicy {
         if (!isApplicable(discountContext)) {
             return 0;
         }
-        return (int) discountContext.getOrderItems().getItemsByType(MenuType.MAIN)
+        return discountContext.getOrderItems().getItemsByType(MenuType.MAIN)
             .values()
             .stream()
             .mapToInt(Integer::intValue)
